+++ jerry-core/parser/js/js-parser-util.c
@@ -117,7 +117,7 @@ parser_flush_cbc (parser_context_t *context_p) /**< context */
 
     JERRY_ASSERT (opcode < CBC_EXT_END);
     flags = cbc_ext_flags[opcode];
-    parser_emit_two_bytes (context_p, CBC_EXT_OPCODE, opcode);
+    parser_emit_two_bytes (context_p, CBC_EXT_OPCODE, (uint8_t) opcode);
     context_p->byte_code_size += 2;
   }
